requirements.txt
absl-py==1.0.0
astunparse==1.6.3
boto3==1.13.10
botocore==1.16.10
Brotli==1.0.7
cachetools==5.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.11.0
dash==1.12.0
dash-core-components==1.10.0
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
docutils==0.15.2
Flask==1.1.2
Flask-Caching==1.8.0
Flask-Compress==1.5.0
flatbuffers==2.0
fonttools==4.29.1
future==0.18.2
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==20.0.4
h5py==3.6.0
idna==2.9
imageio==2.14.1
importlib-metadata==4.10.1
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==1.1.1
matplotlib==3.5.1
networkx==2.6.3
numpy==1.22.2
oauthlib==3.2.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.0.3
Pillow==9.0.1
plotly==4.7.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
PyWavelets==1.2.0
redis==3.5.1
requests==2.23.0
requests-oauthlib==1.3.1
retrying==1.3.3
rsa==4.8
s3transfer==0.3.3
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.4.1
seaborn==0.11.2
six==1.14.0
sklearn==0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2022.2.2
typing_extensions==4.0.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.13.3
zipp==3.7.0
